Move SSL renewal handling to standalone command.
