+++ config/events.php
@@ -138,7 +138,6 @@ return function (CallableEventDispatcherInterface $dispatcher) {
                 App\Sync\Task\CleanupStorageTask::class,
                 App\Sync\Task\MoveBroadcastsTask::class,
                 App\Sync\Task\ReactivateStreamerTask::class,
-                App\Sync\Task\ReloadFrontendAfterSslChangeTask::class,
                 App\Sync\Task\RotateLogsTask::class,
                 App\Sync\Task\RunAnalyticsTask::class,
                 App\Sync\Task\RunAutomatedAssignmentTask::class,
