+++ util/docker/web/scripts/on_ssl_renewal
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+
+sv restart nginx
+
+azuracast_cli azuracast:internal:on-ssl-renewal
